IFile Integration and Concrete ResourceCost classes
I have integrated the IFile interface with the CRUD class which should allow us to read from any file and write to any file once you make a concrete class for a specific file type. Should also scale to databases if we end up doing one. Also made a number of concrete classes for the Resources class (now called ResourceCost) and Structure now has an arraylist of ResourceCosts. Didn't want to completely go all in on the ResourceCost integration until we talk tommorow with JJ to make sure this is the structure he wants