ResourceCost added to main branch
So I think the ResourceCost thing I was working on the last day should qualify for the state pattern - "State is a behavioral design pattern that allows an object to alter its behavior when its internal state changes". So the idea is that instead of a static way to calculate the cost of a specific concrete building, a building could have many different ways to calculate based on whatever factor we want to - i.e. houses with students could recieve a discount, or factories that would use eccessive water would be charged more for that water on a per litre basis compared to a regular house. It's a bit unrealistic but if JJ's okay with it, we have another design pattern down. I'll try to explain it better tommorow at the metting.